[216] Move shared tabular preparation into synthefy (#423)
* refactor: centralize tabular preparation in synthefy

* test: lock tabular preparation ownership

* test: preserve pre-move featurizer pickles
