+++ docs/architecture/0001-consolidate-synthefy-source-tree-ci-gates.json
@@ -80,6 +80,20 @@
       "retirement_action": "Delete the private duplicate implementation and exact-submodule fallback, then retain the canonical synthefy.text_features identity, pickle, and optional-extra coverage."
     },
     {
+      "id": "legacy-6.3-tabular-featurizer-fallback",
+      "lifecycle": "migration_scaffolding",
+      "status": "active",
+      "paths": [
+        "src/synthefy_nori/_legacy_featurize.py",
+        "src/synthefy_nori/featurize.py",
+        "tests/test_featurize.py",
+        "tests/test_client_matches_api.py",
+        ".github/workflows/client-sync.yml"
+      ],
+      "retirement_event": "The released-client compatibility floor and client-sync lane are deliberately advanced to synthefy >=7.0.0 after the consolidated client release.",
+      "retirement_action": "Delete the private duplicate implementation and exact-submodule fallback, then retain canonical synthefy.featurize identity, behavior, and client-only artifact coverage."
+    },
+    {
       "id": "phase-2-open-gap-observation",
       "lifecycle": "migration_scaffolding",
       "status": "active",
